src/client/app/ng-magic-mirror.ts
import {Component} from 'angular2/core';
import {RouteConfig, ROUTER_DIRECTIVES, ROUTER_PROVIDERS} from 'angular2/router';

@Component({
  moduleId: __moduleName,
  selector: 'ng-magic-mirror-app',
  providers: [ROUTER_PROVIDERS],
  templateUrl: 'ng-magic-mirror.html',
  styleUrls: ['ng-magic-mirror.css'],
  directives: [ROUTER_DIRECTIVES],
  pipes: []
})
@RouteConfig([
])
export class NgMagicMirrorApp {
  defaultMeaning: number = 42;

  meaningOfLife(meaning?: number) {
    return `The meaning of life is ${meaning || this.defaultMeaning}`;
  }
}
